Star Wars theme played by Piezo and Mbed
connected mbed p21 to one leg of the piezo transducer, the other leg was connected to ground.  Then, I researched the corresponding frequencies for each of the notes in the main part of the legendary Star Wars theme. Mbed has a .period function, so I used the relationship between frequency and period to convert the frequency into a period. Then, I put them in the correct order, gave them the correct duration, and boom.